test: rename tests and update imports
