+++ packages/amplify-category-api/src/__tests__/provider-utils/awscloudformation/utils/resolver-config-to-conflict-resolution-bi-di-mapper.test.ts
@@ -2,7 +2,7 @@ import { ConflictResolution } from 'amplify-headless-interface';
 import {
   conflictResolutionToResolverConfig,
   resolverConfigToConflictResolution,
-} from '../../../../provider-utils/awscloudformation/utils/resolverConfigToConflictResolutionBiDiMapper';
+} from '../../../../provider-utils/awscloudformation/utils/resolver-config-to-conflict-resolution-bi-di-mapper';
 import { ResolverConfig, ConflictHandlerType } from 'graphql-transformer-core';
 
 describe('transform ConflictResolution to ResolverConfig', () => {
